build: table component state
