+++ src/components/UserManagement/TableComponent/state.js
@@ -1,0 +1,32 @@
+export const useTableComponentState=()=>{
+
+  const descendingComparator=(a, b, orderBy)=>{
+    if (b[orderBy] < a[orderBy]) {
+      return -1
+    }
+    if (b[orderBy] > a[orderBy]) {
+      return 1
+    }
+    return 0
+  }
+  const getComparator=(order, orderBy) => {
+    return order === 'desc'
+      ? (a, b) => descendingComparator(a, b, orderBy)
+      : (a, b) => -descendingComparator(a, b, orderBy)
+  }
+  const stableSort=(array, comparator) =>{
+    const stabilizedThis = array.map((el, index) => [el, index])
+    stabilizedThis.sort((a, b) => {
+      const order = comparator(a[0], b[0])
+      if (order !== 0) return order
+      return a[1] - b[1]
+    })
+    return stabilizedThis.map((el) => el[0])
+
+  }
+  return{
+    getComparator,
+    stableSort
+  }
+
+}
